Refactor App.vue and views
